Import Garden Linux source packages (#7)
Ingest source packages from Garden Linux. This is required for glvd to provide accurate information.

Note: This won't work for existing releases, it should be in for releases newer than 1443.15 and 1592.2.

Only import data for Garden Linux today because for now this is the only suite that provides a source repo.