Ingest Kernel CVE information (#22)
We need special treatment for kernel CVE information because the data provided by NVD is not enough for us.
NVD does not maintain data on fixed CVEs in LTS kernels, but for us this is very relevant.
kernel.org provides data on that instead, so we need to take this into account additionally.

Part of the solution for https://github.com/gardenlinux/glvd/issues/122