Rework DB Schema (#30)
The Database schema is created as a side effect of the ingestion process so far.
This leads to some in-transparencies with regards to the schema.
This PR moves the responsibility to create the schema out of the python code, and clearly defines the schema in an SQL file so it is independent of the ORM used.

Also, this PR adds a `gardenlinux_version` field that will relieve us from the need to join the `dist_cpe` table in many cases. This makes the schema less generic, and does not follow the textbook design for relational databases, but I think this is a justified shortcut with little downsides in a database that is not meant as a generic Linux Distribution security tracker, but rather a Vulnerability Database for Garden Linux.

Adds schema migrations based on https://nalanj.dev/posts/minimalist-postgresql-migrations/

Fixes https://github.com/gardenlinux/glvd/issues/151