Take minor version into account for debsec_vulnerable calculation (#49)
Mark CVEs as vulnerable when the 'fixed version' has a different minor version from the version we have.
This might create false positives, but without this we get a lot of false negatives, for example in glibc where the minor version is updated often.
Comparing glibc version 2.41-3 to 2.39-17 makes no sense in this case as there is no way to tell if the first version includes any CVE fix provided for the latter.

To reduce the number of false positives,  we parse the changelogs for mentions of CVEs, marking those as 'resolved' that are mentioned in a changelog. This is not perfect, but it should be a good enough heuristic.

Further false positives need to be addressed via triage.

Fixes https://github.com/gardenlinux/glvd/issues/149