Filter kernel cves from default cve view (#75)
This is a regression that was introduced in https://github.com/gardenlinux/glvd-data-ingestion/pull/67

We have special handling for kernel cves because the 'fixed' information
from the debian security tracker is not useful for us since we follow
lts kernel versions and maintain our own builds.

This change must have been done by accident.