Allow upgrading schema to latest
This PR rewrite the migrate and migrate-all scripts in python. This is done to make it easier to have more logic in them.

The need to do it in this case is that we want to implement a release process for GLVD.

It should after applying this change be possible to run the migrate-all script in a released version and thus upgrade an older version of a db to the newer version.