Import source manifests for already released versions (#88)
Future Garden Linux releases will have a source manifest as part of the release. For existing releases, this is not going to be the case, so we import those once to enable the new image-view feature in GLVD also for those releases.

This PR is a prerequisite for #86 

Part of https://github.com/gardenlinux/glvd/issues/189